+++ src/compounds/product-table/src/components/cell-image.tsx
@@ -4,14 +4,22 @@ import { jsx } from 'theme-ui'
 
 import CellBase from './cell-base'
 
-export const ProductTableCellImage: React.FC = ({ children }) => (
+export interface Props extends React.HTMLAttributes<HTMLDivElement> {
+  headerImage?: boolean
+}
+
+export const ProductTableCellImage: React.FC<Props> = ({
+  children,
+  headerImage
+}) => (
   <CellBase
-    mobileOrder={0}
+    mobileOrder={!headerImage ? 0 : 2}
     sx={{
       display: 'flex',
       alignItems: 'center',
       justifyContent: 'center',
-      variant: 'compounds.product-table.cellImage.main'
+      variant: `compounds.product-table.${headerImage &&
+        'variants.redesign.'}cellImage.main`
     }}
   >
     <div
@@ -19,8 +27,10 @@ export const ProductTableCellImage: React.FC = ({ children }) => (
         position: 'relative',
         height: '100%',
         width: '100%',
-        minHeight: 75,
-        minWidth: 100,
+        minHeight: headerImage ? 90 : 75,
+        minWidth: headerImage ? 120 : 100,
+        variant: `compounds.product-table.${headerImage &&
+          'variants.redesign.'}cellImage.imageWrapper`,
         img: {
           position: 'absolute',
           top: 0,
@@ -30,7 +40,9 @@ export const ProductTableCellImage: React.FC = ({ children }) => (
           maxWidth: '100%',
           maxHeight: '100%',
           margin: 'auto',
-          objectFit: 'contain'
+          objectFit: 'contain',
+          variant: `compounds.product-table.${headerImage &&
+            'variants.redesign.'}cellImage.image`
         }
       }}
     >