+++ src/elements/accordion/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.7](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.accordion@0.9.6...@uswitch/trustyle.accordion@0.9.7) (2020-10-28)
+
+
+### Bug Fixes
+
+* add redesign variant styling to accordion ([101f5a7](https://github.com/uswitch/trustyle/commit/101f5a7))
+
+
+
+
+
 ## [0.9.6](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.accordion@0.9.5...@uswitch/trustyle.accordion@0.9.6) (2020-10-28)
 
 **Note:** Version bump only for package @uswitch/trustyle.accordion